--- src/main/java/org/devcourse/resumeme/business/resume/repository/CareerRepository.java
@@ -1,8 +1,0 @@
-package org.devcourse.resumeme.business.resume.repository;
-
-import org.devcourse.resumeme.business.resume.domain.Career;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CareerRepository extends JpaRepository<Career, Long> {
-
-}
